+++ Cargo.toml
@@ -6,21 +6,15 @@ license = "Apache-2.0"
 documentation = "https://jgrillo.github.io/forecast-rs/"
 repository = "https://github.com/jgrillo/forecast-rs"
 keywords = ["forecast", "weather", "darksky"]
-version = "0.1.0"
+version = "0.2.0"
 authors = ["Jesse C. Grillo <jesse.grillo@gmail.com>"]
-build = "build.rs"
 
 [features]
-default = ["serde_codegen"]
-unstable = ["serde_derive"]
 integration = []
 
-[build-dependencies]
-serde_codegen = { version = "0.8.10", optional = true }
-
 [dependencies]
-serde = "0.8.9"
-serde_derive = { version = "0.8.9", optional = true }
-serde_json = "0.8.2"
-hyper = "0.9.12"
-itertools = "0.5.4"
+serde = "0.9"
+serde_derive = "0.9"
+serde_json = "0.9"
+reqwest = "0.4"
+itertools = "0.5"
